refactor view to ui
