+++ lua/hyper/core.lua
@@ -1,7 +1,7 @@
-local RequestScreen = require("hyper.view.screens.request")
-local HistoryScreen = require("hyper.view.screens.history")
-local VariablesScreen = require("hyper.view.screens.variables")
-local CollectionScreen = require("hyper.view.screens.collections")
+local RequestScreen = require("hyper.screens.request")
+local HistoryScreen = require("hyper.screens.history")
+local VariablesScreen = require("hyper.screens.variables")
+local CollectionScreen = require("hyper.screens.collections")
 local State = require("hyper.state")
 local Envs = require("hyper.envs")
 local Http = require("hyper.utils.http")
