+++ lua/hyper/screens/collections.lua
@@ -1,8 +1,8 @@
-local Float = require("hyper.view.float")
-local Selector = require("hyper.view.selector")
-local Screen = require("hyper.view.screen")
+local Float = require("hyper.ui.float")
+local Selector = require("hyper.ui.selector")
+local Screen = require("hyper.ui.screen")
+local Text = require("hyper.ui.text")
 local Collections = require("hyper.collections")
-local Text = require("hyper.view.text")
 local Ui = require("hyper.utils.ui")
 local Http = require("hyper.utils.http")
 local hyper = require("hyper")
