+++ lua/hyper/screens/history.lua
@@ -1,10 +1,10 @@
-local Float = require("hyper.view.float")
-local Selector = require("hyper.view.selector")
-local Screen = require("hyper.view.screen")
+local Float = require("hyper.ui.float")
+local Selector = require("hyper.ui.selector")
+local Screen = require("hyper.ui.screen")
 local Ui = require("hyper.utils.ui")
 local Http = require("hyper.utils.http")
 local History = require("hyper.history")
-local Text = require("hyper.view.text")
+local Text = require("hyper.ui.text")
 local hyper = require("hyper")
 
 local width, height, row, col = Ui.get_viewbox()
