+++ lua/hyper/ui/file.lua
@@ -1,4 +1,4 @@
-local Float = require("hyper.view.float")
+local Float = require("hyper.ui.float")
 
 local File = {}
 File.__index = File
